Added target mismatch detection earlier in code
The quantum neural networks do not like seeing fewer targets in their input than they expect in their output. Doing so results in an error, which can be very annoying after running a lengthy effective dimension computation. This pushes that error earlier on in the code so it doesn't waste so much time. (Current fix for the error is just to rerun the code/change your data settings to try and get all the targets in the first batch.)