Added comment on new dynamic block input format
The dynamic rotation block factory no longer just takes an input (for now). It requires an array of ints that specifies what types of block should be placed at each qubit. The old behaviour can be replicated with a list comprehension of the desired sized filled with your favourite number greater than 1.